travis: dist: precise is being decommissioned.
https://blog.travis-ci.com/2017-08-31-trusty-as-default-status

So we need to switch to dist: trusty which only has openjdk7 instead of
oraclejdk7. https://github.com/travis-ci/travis-ci/issues/8439.

openjdk7 does not include the necessary ciphers to install gradle from
the wrapper. So we must use oraclejdk8 to install gradle from the
wrapper.

trusty also sets _JAVA_OPTIONS which upsets some tests, so we must make
sure it is unset.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>